Split View and convert QueryBuilder.tsx to typescript
