+++ frontend/src/metabase/query_builder/components/view/View/NativeQueryRightSidebar/index.ts
@@ -1,0 +1,1 @@
+export * from "./NativeQueryRightSidebar";
