+++ frontend/src/metabase/query_builder/components/view/View/StructuredQueryRightSidebar/index.ts
@@ -1,0 +1,1 @@
+export * from "./StructuredQueryRightSidebar";
