+++ frontend/src/metabase/query_builder/components/view/View/ViewLeftSidebarContainer/index.ts
@@ -1,0 +1,1 @@
+export * from "./ViewLeftSidebarContainer";
