+++ frontend/src/metabase/query_builder/components/view/View/ViewNativeQueryEditor/index.ts
@@ -1,0 +1,1 @@
+export * from "./ViewNativeQueryEditor";
