+++ frontend/src/metabase/query_builder/components/view/View/index.ts
@@ -1,0 +1,1 @@
+export { View } from "./View";
